Move annotated components back to the tests
